requirements.txt
astroid==3.1.0
dill==0.3.8
isort==5.13.2
mccabe==0.7.0
numpy==1.26.4
pandas==2.2.0
platformdirs==4.2.0
psycopg2==2.9.9
pyarrow==15.0.2
pylint==3.1.0
python-dateutil==2.8.2
pytz==2024.1
six==1.16.0
tomlkit==0.12.4
tqdm==4.66.2
tzdata==2024.1
